x64構成を作ってみました。 あとはBoost関連のライブラリなんですが、ビルドの仕方がわからなかったです。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5187 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Graphics/GRMesh.h
@@ -115,7 +115,7 @@ public:
 	Vec2f*		GetTexCoords(){ return &texCoords[0]; }
 	GRMeshFace*	GetFaces(){ return &faces[0]; }
 	GRMeshFace*	GetFaceNormals(){ return &faceNormals[0]; }
-	unsigned*	GetTriangles(){ return &triFaces[0]; }
+	size_t*		GetTriangles(){ return &triFaces[0]; }
 	int*		GetMaterialIndices(){ return &materialList[0]; }
 	void		SwitchCoordinate();
